[Refactor] 라우팅 오류 해결
