+++ src/components/layout/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router';
 import styled from 'styled-components';
-import Header from '../domain/Header';
+import Header from '../Header';
 
 const Layout = () => {
   return (
