+++ src/routes/router.tsx
@@ -1,7 +1,7 @@
-import Scraps from '@/components/domain/my-page/Scraps';
-import UserContents from '@/components/domain/my-page/UserContents';
-import UserPosts from '@/components/domain/my-page/UserPosts';
 import Layout from '@/components/layout/Layout';
+import Scraps from '@/components/my-page/Scraps';
+import UserContents from '@/components/my-page/UserContents';
+import UserPosts from '@/components/my-page/UserPosts';
 import HomePage from '@/pages/HomePage';
 import MyPage from '@/pages/MyPage';
